Fix pylint issues.
